Issue: "port" module on PIC16F8x (SOLVED)

+++ ports/Microchip/PIC16F-generic/api/port/read.c.v
@@ -10,6 +10,6 @@ module port
 @[inline]
 pub fn read(name u8) u8 {
 	unsafe {
-		return *(&C.PORTA + name)
+		return u8(*(&C.PORTA + name))
 	}	
 }
